Entrega1/requirements.txt
psycopg2-binary
python-dotenv
mediapipe
opencv-python
pandas
pydrive2
numpy
